move to safety_lateral (TODO: fix order and formatting)
